corrects mistake when hydro_year not equal to '01/10'
The argument 'hydro_year' in the call of the 'calculator' function takes a default value of '01/10'. In order to make EFlowCalc more flexible, namely usable in the Southern hemisphere, the user can set a different day and month for the start of the hydrological year. However, there were still some datetime strings still hard coded using the 1st of October as the starting date for a hydrological year.

This commit fixes this, and now the 'hydro_year' change is effectively honoured by EFlowCalc, which was not the case in any previous version.

This commit also includes a check on the format of the 'hydro_year' argument to make sure it is syntactically and semantically correct.