adds return condition to send back array in original orientation
The function 'calculator' transposes a copy of the input array before proceeding with the calculation of the SFCs, but the array returned was not transposed back, to return an array with matching orientation compared with the input array.

This is improved by adding a conditional on the return, to transpose back if necessary.