deals with zero divide in DL19 and mean of empty slice in FL3
For DL19 (Variability in annual number of zero flow days), very often the denominator (the mean) equals to zero, which returns a NaN and raises a warning. To avoid this, true_divide with condition 'where' is used.

For FL3 (Mean annual low flow pulse count) can often produce an empty slice when no low flow pulse is found, which raise a warning and returns a NaN value. To avoid this, NaN values are replaced by zeros when necessary.