corrects mistake in 'count_reversal' function
In the block of code that intends to replace null differences between consecutive flows, there was a mistake in the index slicing for the second axis where to replace values in place when difference is null. It was using the axis 0 from the output of np.nonzero to slice on the axis 1 of the diff array.

For simulated flows, it is highly unlikely that null differences occur, however it is much more likely to occur on observed flows because of the low level of accuracy of the reported mesurements. This can result in numerical mistakes, and should most of the time have raised an IndexError, unless the time dimension is shorter than the number of simulations (which is often only 1 for observed timeseries). The SFCs potentially impacted were RA8 and RA9.

The error is now fixed.