improves 'count_reversal' to deal with special case
If the null difference in flows happens to be at the very beginning of the array 'diff', it has no previous value to forward replace the null value (intentionally set to NaN beforehand), so it will replicate the NaN value(s).

Mirroring what is done for forward replacing using maximum.accumulate (from start to end of axis 0), a backward replacement is used with minimum.accumulate (from end to start of axis 0) to deal with NaNs at the very beginning of the array 'diff'.

Because the input flow array 'arr' is checked for NaN values in the call of 'calculator', there is no reason for a NaN value to be present in 'diff' unless it was intentionally set as such by 'count_reversal' when null differences where found. As a result, it is safe to backward replace, because the only NaN values remaining can only be at the very beginning, all other NaN values have been replaced by the first block of code containing maximum.accumulate.